requirements.txt
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
notebooklm-py==0.1.4
patchright==1.57.2
playwright==1.57.0
pyaes==1.6.1
pyasn1==0.6.1
pyee==13.0.0
Pygments==2.19.2
python-dotenv==1.2.1
python-socks==2.8.0
rich==14.2.0
rsa==4.9.1
Telethon==1.42.0
typing_extensions==4.15.0
